+++ lib/global/cursor/updateCursorStyle.ts
@@ -10,7 +10,12 @@ const documentToStyleMap = new WeakMap<
 >();
 
 export function updateCursorStyle(ownerDocument: Document) {
-  if (ownerDocument.defaultView === null) {
+  // NOTE undefined is not technically a valid value but it has been reported that it is present in some environments (Vite HMR?)
+  // See github.com/bvaughn/react-resizable-panels/issues/559
+  if (
+    ownerDocument.defaultView === null ||
+    ownerDocument.defaultView === undefined
+  ) {
     return;
   }